refactor: update schematic graphic renderer imports
