+++ tests/sch/schematic-graphic3.test.ts
@@ -1,5 +1,5 @@
 import { expect, test } from "bun:test"
-import { createSvgObjectFromSchematicGraphic } from "lib/sch/svg-object-fns/create-svg-object-from-sch-graphic"
+import { createSvgObjectFromSchematicGraphic } from "lib/sch/svg-object-fns/create-svg-object-from-schematic-graphic"
 import {
   decodeSvgDataUrl,
   getEmbeddedImage,
