+++ tests/sch/schematic-graphic6.test.ts
@@ -1,5 +1,5 @@
 import { expect, test } from "bun:test"
-import { createSvgObjectFromSchematicGraphic } from "lib/sch/svg-object-fns/create-svg-object-from-sch-graphic"
+import { createSvgObjectFromSchematicGraphic } from "lib/sch/svg-object-fns/create-svg-object-from-schematic-graphic"
 import { getEmbeddedImage, svgAsset } from "./schematic-graphic-test-helpers"
 
 test("passes a canonical base64 SVG asset through without decoding it", () => {
